Mutualize vehicle creation code
